fix(calculator): comprehensive section audit — 18 bugs fixed + perf verified
High:
- BodyFat: US Navy formula now converts cm→inches (was overstating BF% ~2x)
- HalfLife: mg→oz used g/oz factor (1000x too large) — corrected at 4 sites
- CycleCalendarExporter: ICS dates used UTC toISOString (one-day shift for
  UTC+ users) + crash on cleared date input — local-date serialization + guard

Medium:
- Macro: simulation W0 off-by-one (start weight shown after 1 week) + '12-week'
  projection read 13 weeks of change; case-sensitive activity bonus
- BodyFat: clearing an input no longer leaves stale base measurements
- MasterCalculator: goal i18n keys (bulking/cutting/trt/sports) + parseInt→float
- formulaEngine: budget modifier now applied (low/med/high were a no-op)
- GeneticPotential: golden-ratio/structure score was constant (used skeletal
  proportions not user build), ffmi field mislabeled, reset cleared base values
- HalfLife: tooltip broke compound ids with underscores (test_e→'test'),
  dynamic Tailwind grid class not generated, dead >1000mg hazard block revived
- ReadinessQuiz: setScore updater side effects double-advanced questions in
  StrictMode; onComplete prop was never wired
- transformationEngine: goalProgressPct used fat-only loss vs scale target,
  contradicting weeksToIdeal (gauge 100% while 9 weeks left)
- SmartLabReference: category icons looked up by icon name not category id
- useSmartLabReference: setter called inside setState updater (StrictMode)

Low:
- AnimatedNumber rAF cleanup, clipboard guards (quiz + daily IQ), half-life
  stability window fallback for short-compound stacks, bodyfat 'fitness'
  category description (was mapped to athletes)

Tests: navyFormula.spec, formulaEngine.spec, goalProgress consistency;
280/280 pass, tsc/lint clean, build ok. i18n parity verified (1210 keys each).

+++ src/features/calculator/MasterCalculator.tsx
@@ -130,8 +130,7 @@ const MasterCalculator: React.FC<MasterCalculatorProps> = ({ navigateTo }) => {
                                     onClick={() => setState(prev => ({ ...prev, goal: g }))}
                                     className={`py-3 px-4 rounded-xl text-[10px] font-black uppercase tracking-tighter transition-all border-2 ${state.goal === g ? 'bg-gold-500/10 border-gold-500 text-gold-500' : 'bg-zinc-50 dark:bg-zinc-900 border-transparent text-zinc-500 hover:border-zinc-700'}`}
                                 >
-                                    {content.calcSelectGoal ? (content.calcSelectGoal[g as keyof typeof content.calcSelectGoal] || g) : g}
-                                </button>
+                                    {content.calcSelectGoal ? (content.calcSelectGoal[g as keyof typeof content.calcSelectGoal] || g) : g}                                </button>
                             ))}
                         </div>
                     </div>
@@ -198,7 +197,7 @@ const MasterCalculator: React.FC<MasterCalculatorProps> = ({ navigateTo }) => {
                                 type="number"
                                 title="Enter weight"
                                 value={state.weight}
-                                onChange={(e) => setState(prev => ({ ...prev, weight: parseInt(e.target.value) || 0 }))}
+                                onChange={(e) => setState(prev => ({ ...prev, weight: parseFloat(e.target.value) || 0 }))}
                                 className="w-full bg-zinc-50 dark:bg-zinc-900 border border-zinc-800 rounded-xl p-3 text-center font-mono font-black text-lg focus:border-gold-500 outline-none"
                             />
                         </div>
